Improve sync send logic for allowing batching
*Motivation*

Batching is required for achieving high performance messaging. If we want to flip the default behaviour of publishers into batch publishing,
we need to improve current send logic to flush out batches after enqueuing the send requests. so batch settings won't impact sync sends.

*Modification*

- add a flush() method on single partition producer
- break the send logic on single partition producer to `enqueue-flush-wait`
- change multi-partitions producer to use single partition producer
